fastapi/requirements.txt
fastapi
uvicorn
openai
snowflake-connector-python
pydantic
